Module sha::sha1::ops

TODO: docs

Functions

digest
TODO

digest_block
TODO

digest_round_x20

digest_round_x4
This function can be easily implemented with Intel SHA intruction set extensions.

expand_round_x20

expand_round_x4
This function can be easily implemented with Intel SHA intruction set extensions.

pad