Module sha::sha512::ops

TODO: docs

Functions

digest
TODO

digest_block
TODO

digest_round_x16
TODO

digest_round_x4
There are no plans for hardware implementations at this time, but this function can be easily implemented with some kind of SIMD assistance.

expand_round_x16
TODO

expand_round_x4
There are no plans for hardware implementations at this time, but this function can be easily implemented with some kind of SIMD assistance.